Add maps to C4Script
{} creates an empty map, except where a block is expected.
{ key = "value", "key with non-identifier-characters" = 1337, ... } creates a map with according key value-pairs.
in { key = "value", } the last comma is allowed for easier key-value pair additions later on.

The entries of a map can be iterated using for(var key, value in map).
GetKeys(map) and GetValues(map) can be used to get an array of all keys and values of a map.

To access a specific map element the []-operator can be used, e.g. map["foo"].
To access entries with a constant string key, the access-operator . can be used, e.g. map.foo accesses the same entry as map["foo"].
All value types can be used as keys, including arrays and maps.
Though, for the {}-initialization and map.key access, only strings are supported.

By setting an entry to 0, the according key-value pair gets removed from the map.

To stay in sync, the iteration order as well as the order of the result of GetKeys() and GetValues() resembles the order in which the entries have been inserted into the map.

WARNING: When using objects or objects nested inside maps or arrays as keys, the assigned values may become inaccessible when the object is being removed.