Don't explicitly check if a value or reference is needed for array access on strings
If the caller tries to modify the returned value, an error will occur anyway
because the returned value is a string and not a reference.
Otherwise with the explicit check many false positives are detected, because SetNoRef is not used everywhere needed.
Thus, AB_ARRAYA_V vs. AB_ARRAYA_R can only be considered a hint.