Add experimental "preloading" support.
Preloading refers to loading most of the game data in the background while being in the lobby.
Automatic preloading as soon as all data has been loaded can be enabled by setting the config entry General.Preloading.
Preloading can also be triggered manually (without changing the config) with the corresponding button in the resources tab of the lobby.