Reintroduce Ev_HTTP_Response in order to fix a crash due to a race condition
Using he new lambda callback system for C4Network2HTTPClient for C4StarutpNetDlg
has a race condition - the entry might disappear and delete the client while a
callback event is still in C4InteractiveThread's queue, accessing invalid data.
This commit Reintroduce Ev_HTTP_Response to fix this issue - if such an event
remains in the queue, it is simply discarded.