C4Movement: Do not call Hit() with parameters in scripts older than CR
In CE and previous Clonk versions, Hit() was never called with parameters.
This new behaviour of CR causes problems with older scenarios like Clonk
Planet's Research Facility, as it calls Hit with an object argument (or
lets the engine call it without any arguments), which in turn creates
errors as C4Movement calls it with two integers. Therefore, this commit
disables those parameters for object scripts older than Clonk Rage in
which theses parameters got introduced.