Change over to significantly better connections if available
A connection is considered better if the candidate’s average ping is shorter than 2/3 of the current connection’s ping.
This should remedy situations where IPv4, IPv6, TCP or UDP connectivity is significantly weaker than the alternative but the bad one was chosen initially.
In order to have alternative connections available, all possible connections are kept for backup.

Also, addresses retrieved from the puncher are now inserted as fist addresses to circumvent problems where the public address doesn’t fit in the limited array of client addresses in the masterserver reference.