Only pass the material part of the `Landscape.Material` combination to C4MaterialMap::Get (fixes #104)
C4MapCreator will happily accept a material-texture-combination. Looking up
the earth material index, however, will not, which caused `InEarth` and `Nest`
objects to be placed in front of Sky since the Earth material could not be
found.

This commit changes the lookup to only use the material part for scenarios
targeting at least engine version [359]. For older scenarios, a debug message
is logged if `InEarth` or `Nest` aren't empty and the old behavior of setting
`MEarth` to `MNone` is preserved.