Replace Log and LogF with spdlog::logger-based logging
- Log is now a shortcut for LoadResStr. It takes a compile-time string table
key, checks the argument count, formats it and logs the result.
- LogNTr logs its argument for the cases where you don't need translation.
- LogSilent has been replaced with the default logger (spdlog::log), with
various loggers with different names being used for different subsystems,
e.g. the network stack.

Everything logged with Log / LogNTr is always displayed in the GUI, whereas
messages logged via other means are only display if they are a warning,
an error or critical, unless /verbose is passed via the command line.