C4Group: Workaround Linux 6.11 and 6.12 removing EXTBSY for executable files
Linux 6.11 and 6.12 allow writing to running executables, including c4group
extracting to its own c4group binary, corrupting it in the process. While
this will be reverted as it caused regressions, we cannot force our users
to use a different kernel version. As the normal extraction process already
renames a temporary file over the executable instead of writing to it,
change the already questionable file reservation process to be exclusive -
after all, if the file already exists, it's already 'reserved'.